+++ app/src/main/java/one/mixin/android/job/TranscriptAttachmentMigrationJob.kt
@@ -5,8 +5,10 @@ import com.birbit.android.jobqueue.Params
 import kotlinx.coroutines.runBlocking
 import one.mixin.android.Constants
 import one.mixin.android.MixinApplication
+import one.mixin.android.extension.getMediaPath
 import one.mixin.android.extension.getTranscriptDirPath
 import one.mixin.android.extension.hasWritePermission
+import timber.log.Timber
 import java.nio.file.Files
 
 class TranscriptAttachmentMigrationJob : BaseJob(Params(PRIORITY_LOWER).groupBy(GROUP_ID).persist()) {
@@ -17,11 +19,20 @@ class TranscriptAttachmentMigrationJob : BaseJob(Params(PRIORITY_LOWER).groupBy(
     override fun onRun() = runBlocking {
         if (!hasWritePermission()) return@runBlocking
         val oldDir = MixinApplication.get().applicationContext.getTranscriptDirPath(true)
-        val newDir = MixinApplication.get().applicationContext.getTranscriptDirPath(false)
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-            Files.move(oldDir.toPath(), newDir.toPath())
+        if (oldDir.exists()) {
+            val newDir = MixinApplication.get().applicationContext.getTranscriptDirPath(false)
+            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
+                Files.move(oldDir.toPath(), newDir.toPath())
+            } else {
+                oldDir.renameTo(newDir)
+            }
+            Timber.d("Transcript attachment migration ${oldDir.absolutePath} ${newDir.absolutePath}")
         } else {
-            oldDir.renameTo(newDir)
+            Timber.d("Transcript attachment migration old not exists")
+        }
+
+        if (propertyDao.findValueByKey(Constants.Account.Migration.PREF_MIGRATION_ATTACHMENT)?.toBoolean() != true) {
+            MixinApplication.appContext.getMediaPath(true)?.deleteRecursively()
         }
         propertyDao.updateValueByKey(Constants.Account.Migration.PREF_MIGRATION_TRANSCRIPT_ATTACHMENT, false.toString())
     }
